changed Integer.parseInt to Long.parseLong

+++ Main.java
@@ -4,8 +4,8 @@ public class Main {
 
     public static void main(String[] args) {
         if (args.length == 2) {
-            long initialNumberA = Integer.parseInt(args[0]);
-            long initialNumberB = Integer.parseInt(args[1]);
+            long initialNumberA = Long.parseLong(args[0]);
+            long initialNumberB = Long.parseLong(args[1]);
             Generator generator = new Generator();
             System.out.println(generator.countMatches(initialNumberA, initialNumberB));
             } else {
